pyproject.toml
[project]
name = "real-agents-workshop"
version = "0.1.0"
description = "Build & Optimise AI Agents That Survive Production"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai",
    "openai",
    "langfuse",
    "notebook",
    "ipykernel",
    "python-dotenv",
    "pytest",
    "duckdb",
    "pandas>=3.0.1",
    "openinference-instrumentation-dspy"
]
